pyproject.toml
[project]
name = "code-review-agent"
version = "0.1.0"
description = "Autonomous GitHub PR code review agent powered by Claude"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.49.0",
    "PyGitHub>=2.3.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "ruff>=0.4.0",
    "bandit>=1.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
review = "agent.agent:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
